Moved DigitalOceanAccount and DigitalOceanDomain into domain folder in the DigitalOcean Provider. Finished off the typeconverters for DigitalOcean. Added all DigitalOcean provider services to DI config.
