+++ src/DDNSUpdate/Application/Providers/DigitalOcean/Converters/DNSRecordToDigitalOceanCreateDomainRecordRequestConverter.cs
@@ -6,9 +6,17 @@ namespace DDNSUpdate.Application.Providers.DigitalOcean.Converters
 {
     public class DNSRecordToDigitalOceanCreateDomainRecordRequestConverter : ITypeConverter<DNSRecord, DigitalOceanCreateDomainRecordRequest>
     {
-        public DigitalOceanCreateDomainRecordRequest Convert(DNSRecord record, DigitalOceanCreateDomainRecordRequest request, ResolutionContext context)
+        public DigitalOceanCreateDomainRecordRequest Convert(DNSRecord record, DigitalOceanCreateDomainRecordRequest? request, ResolutionContext context)
         {
             request ??= new DigitalOceanCreateDomainRecordRequest();
+            request.Data = record.Data;
+            request.Flags = record.Flags;
+            request.Name = request.Name;
+            request.Port = record.Port;
+            request.Priority = record.Priority;
+            request.Tag = record.Tag;
+            request.Ttl = record.TTL;
+            request.Weight = record.Weight;
             return request;
         }
     }
