+++ src/DDNSUpdate/Application/Providers/DigitalOcean/DigitalOceanClient.cs
@@ -3,6 +3,7 @@ using System.Net;
 using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using DDNSUpdate.Application.Providers.DigitalOcean.Domain;
 using DDNSUpdate.Application.Providers.DigitalOcean.Requests;
 using DDNSUpdate.Application.Providers.DigitalOcean.Responses;
 using FluentResults;
