+++ src/DDNSUpdate/Application/Providers/DigitalOcean/DigitalOceanConfiguration.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using DDNSUpdate.Application.Providers.DigitalOcean.Domain;
 
 namespace DDNSUpdate.Application.Providers.DigitalOcean
 {
