+++ src/DDNSUpdate/Application/Providers/DigitalOcean/DigitalOceanDNSRecordReader.cs
@@ -2,6 +2,7 @@
 using System.Threading;
 using System.Threading.Tasks;
 using AutoMapper;
+using DDNSUpdate.Application.Providers.DigitalOcean.Domain;
 using DDNSUpdate.Application.Providers.DigitalOcean.Responses;
 using DDNSUpdate.Domain;
 using FluentResults;
