+++ src/DDNSUpdate/Application/Providers/DigitalOcean/DigitalOceanServiceProfile.cs
@@ -9,7 +9,12 @@ namespace DDNSUpdate.Application.Providers.DigitalOcean
         {
             context.Services.AddHttpClient<IDigitalOceanClient, DigitalOceanClient>();
 
+            context.Services.AddScoped<IDigitalOceanAccountProcessor, DigitalOceanAccountProcessor>();
+            context.Services.AddScoped<IDigitalOceanClient, DigitalOceanClient>();
             context.Services.AddScoped<IDDNSService, DigitalOceanDDNSService>();
+            context.Services.AddScoped<IDigitalOceanDNSRecordCreator, DigitalOceanDNSRecordCreator>();
+            context.Services.AddScoped<IDigitalOceanDNSRecordReader, DigitalOceanDNSRecordReader>();
+            context.Services.AddScoped<IDigitalOceanDNSRecordUpdater, DigitalOceanDNSRecordUpdater>();
             context.Services.AddScoped<IDigitalOceanDomainProcessor, DigitalOceanDomainProcessor>();
         }
     }
