+++ src/DDNSUpdate/Application/Providers/DigitalOcean/IDigitalOceanClient.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using DDNSUpdate.Application.Providers.DigitalOcean.Domain;
 using DDNSUpdate.Application.Providers.DigitalOcean.Requests;
 using DDNSUpdate.Application.Providers.DigitalOcean.Responses;
 using FluentResults;
