+++ src/DDNSUpdate/Application/Providers/DigitalOcean/IDigitalOceanDomainProcessor.cs
@@ -1,5 +1,6 @@
 ﻿using System.Threading;
 using System.Threading.Tasks;
+using DDNSUpdate.Application.Providers.DigitalOcean.Domain;
 using DDNSUpdate.Domain;
 using FluentResults;
